Move security related test cases into a its own file.
